% --- Source file:HighProbability-internal.Rd ---
\name{HighProbability-internal}
\alias{get.marginal.probability.tolerance}
\title{ For programmers only. (Not to be called by the user.) }
\description{
	Assign 0.005 for tolerance
}
\usage{
	get.marginal.probability.tolerance()
}
\details{
	None.}
\value{
  Return value 0.005 for tolerance.
}
\references{
	 }
\author{ David R. Bickel (\email{DavidBickel.66846716@bloglines.com}, \url{http://www.davidbickel.com}),  Zahra Montazeri (\email{zahra@math.carleton.ca})  }
\seealso{\code{\link{marginal.probability}}}
\keyword{ htest}